Added clean_data() and clean_file() functions
The original clean_data() function only worked on files. This functionally has now been split. Using clean_data() removes points with low probability and points with zero values from a given data frame. By using clean_file(), which is a wrapper for clean_data(), a file path can be input to the function so that a CSV file can be cleaned without manually loading/saving it.