file_velocity() is a wrapper for velocity_x() and velocity_y()
Can be directly used on CSV files so that data need not be loaded into R first. Simply pass a filename/path and the function creates two output CSV files in the same directory containing the velocity data for the x and y axis computed using velocity_x() and velocity_y().