Imputation of 0 values added
Previously, 0 values were only imputed when the immedately adjacent values were not 0 either. This leads to weird behavior when computing velocity b/c the 0 will be part of the computation and result in huge differences as points supposedly "jump around on the screen". Hence all 0 values in a column will now automatically be replaced with the mean of the closest non-zero values.